+++ test/config.test.js
@@ -54,6 +54,9 @@ describe("Config tests", function () {
         );
         assert(shouldDoInterceptionBasedOnUrl("localhost:3000", "localhost:3000", undefined));
         assert(shouldDoInterceptionBasedOnUrl("https://localhost:3000", "https://localhost:3000", undefined));
+        assert(shouldDoInterceptionBasedOnUrl("https://localhost:3000", "https://localhost:3001", undefined));
+        assert(shouldDoInterceptionBasedOnUrl("http://localhost:3000", "http://localhost:3001", undefined));
+        assert(shouldDoInterceptionBasedOnUrl("localhost:3000", "localhost:3001", undefined));
         assert(shouldDoInterceptionBasedOnUrl("http://localhost:3000", "http://localhost:3000", undefined));
         assert(shouldDoInterceptionBasedOnUrl("localhost:3000", "https://localhost:3000", undefined));
         assert(shouldDoInterceptionBasedOnUrl("localhost", "https://localhost", undefined));
@@ -71,14 +74,22 @@ describe("Config tests", function () {
         assert(shouldDoInterceptionBasedOnUrl("api.example.com", "", "api.example.com"));
         assert(shouldDoInterceptionBasedOnUrl("http://api.example.com", "", "http://api.example.com"));
         assert(shouldDoInterceptionBasedOnUrl("api.example.com", "", ".example.com"));
+        assert(shouldDoInterceptionBasedOnUrl("api.example.com", "", "example.com"));
         assert(shouldDoInterceptionBasedOnUrl("https://api.example.com", "", "http://api.example.com"));
         assert(shouldDoInterceptionBasedOnUrl("https://api.example.com", "", "https://api.example.com"));
+        assert(shouldDoInterceptionBasedOnUrl("https://sub.api.example.com", "", ".sub.api.example.com"));
+        assert(shouldDoInterceptionBasedOnUrl("https://sub.api.example.com", "", "sub.api.example.com"));
         assert(shouldDoInterceptionBasedOnUrl("https://sub.api.example.com", "", ".api.example.com"));
+        assert(shouldDoInterceptionBasedOnUrl("https://sub.api.example.com", "", "api.example.com"));
         assert(shouldDoInterceptionBasedOnUrl("https://sub.api.example.com", "", ".example.com"));
+        assert(shouldDoInterceptionBasedOnUrl("https://sub.api.example.com", "", "example.com"));
         assert(shouldDoInterceptionBasedOnUrl("https://sub.api.example.com:3000", "", ".example.com:3000"));
+        assert(shouldDoInterceptionBasedOnUrl("https://sub.api.example.com:3000", "", "example.com:3000"));
         assert(shouldDoInterceptionBasedOnUrl("https://sub.api.example.com:3000", "", ".example.com"));
+        assert(shouldDoInterceptionBasedOnUrl("https://sub.api.example.com:3000", "", "example.com"));
         assert(shouldDoInterceptionBasedOnUrl("https://sub.api.example.com:3000", "", "https://sub.api.example.com"));
         assert(shouldDoInterceptionBasedOnUrl("https://api.example.com:3000", "", ".api.example.com"));
+        assert(shouldDoInterceptionBasedOnUrl("https://api.example.com:3000", "", "api.example.com"));
         assert(shouldDoInterceptionBasedOnUrl("localhost:3000", "", "localhost:3000"));
         assert(shouldDoInterceptionBasedOnUrl("https://localhost:3000", "", ".localhost:3000"));
         assert(shouldDoInterceptionBasedOnUrl("localhost", "", "localhost"));
@@ -96,31 +107,44 @@ describe("Config tests", function () {
         assert(shouldDoInterceptionBasedOnUrl("https://sub1.api.example.co.uk:3000", "", ".api.example.co.uk"));
         assert(shouldDoInterceptionBasedOnUrl("https://api.example.co.uk:3000", "", ".api.example.co.uk"));
         assert(shouldDoInterceptionBasedOnUrl("https://api.example.co.uk:3000", "", "api.example.co.uk"));
+        assert(shouldDoInterceptionBasedOnUrl("localhost:3000", "localhost:8080", undefined));
+        assert(shouldDoInterceptionBasedOnUrl("localhost:3001", "localhost", undefined));
+        assert(
+            shouldDoInterceptionBasedOnUrl("https://api.example.com:3002", "https://api.example.com:3001", undefined)
+        );
+        assert(shouldDoInterceptionBasedOnUrl(window.location.hostname, "localhost.org:2000", undefined));
+        assert(shouldDoInterceptionBasedOnUrl(window.location.hostname, "localhost", "localhost.org"));
+        assert(shouldDoInterceptionBasedOnUrl("localhost", "localhost", "localhost.org"));
+        assert(shouldDoInterceptionBasedOnUrl("localhost", "", "localhost:8080"));
+        assert(shouldDoInterceptionBasedOnUrl("http://localhost:80", "", "localhost:8080"));
+        assert(shouldDoInterceptionBasedOnUrl("localhost:3000", "", "localhost:8080"));
+        assert(shouldDoInterceptionBasedOnUrl("https://sub.api.example.com:3000", "", ".example.com:3001"));
+        assert(shouldDoInterceptionBasedOnUrl("http://127.0.0.1:3000", "", "https://127.0.0.1:3010"));
 
         // false cases with api
-        assert(!shouldDoInterceptionBasedOnUrl("localhost:3001", "localhost:3000", undefined));
+        assert(!shouldDoInterceptionBasedOnUrl("localhost", "localhost.org"));
+        assert(!shouldDoInterceptionBasedOnUrl("google.com", "localhost.org"));
+        assert(!shouldDoInterceptionBasedOnUrl("http://google.com", "localhost.org"));
+        assert(!shouldDoInterceptionBasedOnUrl("https://google.com", "localhost.org"));
+        assert(!shouldDoInterceptionBasedOnUrl("https://google.com:8080", "localhost.org"));
         assert(!shouldDoInterceptionBasedOnUrl("localhost:3001", "example.com", undefined));
-        assert(!shouldDoInterceptionBasedOnUrl("localhost:3001", "localhost", undefined));
         assert(!shouldDoInterceptionBasedOnUrl("https://example.com", "https://api.example.com", undefined));
         assert(!shouldDoInterceptionBasedOnUrl("https://api.example.com", "https://a.api.example.com", undefined));
+        assert(!shouldDoInterceptionBasedOnUrl("https://api.example.com", "https://a.api.example.com:3000", undefined));
         assert(!shouldDoInterceptionBasedOnUrl("https://api.example.com", "https://example.com", undefined));
         assert(!shouldDoInterceptionBasedOnUrl("https://example.com:3001", "https://api.example.com:3001", undefined));
-        assert(
-            !shouldDoInterceptionBasedOnUrl("https://api.example.com:3002", "https://api.example.com:3001", undefined)
-        );
-        assert(!shouldDoInterceptionBasedOnUrl(window.location.hostname, "localhost.org:2000", undefined));
 
         // false cases with cookieDomain
-        assert(!shouldDoInterceptionBasedOnUrl("https://sub.api.example.com:3000", "", ".example.com:3001"));
-        assert(!shouldDoInterceptionBasedOnUrl("https://sub.api.example.com:3000", "", "example.com"));
+        assert(!shouldDoInterceptionBasedOnUrl("localhost", "", "localhost.org"));
+        assert(!shouldDoInterceptionBasedOnUrl("google.com", "", "localhost.org"));
+        assert(!shouldDoInterceptionBasedOnUrl("http://google.com", "", "localhost.org"));
+        assert(!shouldDoInterceptionBasedOnUrl("https://google.com", "", "localhost.org"));
+        assert(!shouldDoInterceptionBasedOnUrl("https://google.com:8080", "", "localhost.org"));
         assert(!shouldDoInterceptionBasedOnUrl("https://api.example.com:3000", "", ".a.api.example.com"));
         assert(!shouldDoInterceptionBasedOnUrl("https://sub.api.example.com:3000", "", "localhost"));
-        assert(!shouldDoInterceptionBasedOnUrl("http://127.0.0.1:3000", "", "https://127.0.0.1:3010"));
         assert(!shouldDoInterceptionBasedOnUrl(window.location.hostname, "", "localhost"));
         assert(!shouldDoInterceptionBasedOnUrl(window.location.hostname, "", ".localhost"));
         assert(!shouldDoInterceptionBasedOnUrl(window.location.hostname, "", "localhost:2000"));
-        assert(!shouldDoInterceptionBasedOnUrl("https://sub.api.example.co.uk:3000", "", "api.example.co.uk"));
-        assert(!shouldDoInterceptionBasedOnUrl("https://sub.api.example.co.uk", "", "api.example.co.uk"));
 
         // errors in input
         try {
@@ -175,7 +199,7 @@ describe("Config tests", function () {
             normaliseSessionScopeOrThrowError("http://");
             assert(false);
         } catch (err) {
-            assert(err.message === "Please provide a valid sessionTokenFrontendDomain");
+            assert(err.message === "Please provide a valid sessionScope");
         }
     });